Mini-Project-oil-spill-detection/requirements.txt
blinker>=1.8,<2.0
click>=8.0,<9.0
colorama>=0.4,<0.5
filelock>=3.12,<4.0
Flask>=2.3,<4.0
fsspec>=2023.10,<2025.0
itsdangerous>=2.1,<3.0
Jinja2>=3.1,<4.0
MarkupSafe>=2.1,<4.0
mpmath>=1.2,<2.0
networkx>=3.3,<4.0
numpy>=1.23.5,<2.0.0
pillow>=10.0,<12.0
setuptools>=65.5,<80.0
sympy>=1.11,<2.0
torch>=2.0,<3.0
torchvision>=0.15,<1.0
typing_extensions>=4.0,<5.0
Werkzeug>=2.3,<4.0
tensorflow>=2.15,<2.18
